+++ lib/core/options.dart
@@ -78,4 +78,9 @@ class Options {
   ///
   /// @zh: 顶点文字获取器
   String Function(Vertex) textGetter = (Vertex vertex) => '${vertex.id}';
+
+  /// @en: the delay of overlay disappear.
+  ///
+  /// @zh: overlay消失的延迟
+  Duration? panelDelay;
 }