+++ lib/widgets/vertex_component.dart
@@ -47,13 +47,28 @@ class VertexComponent extends ShapeComponent
 
   bool collisionEnable = false;
 
+  String get overlayName => 'vertex${vertex.id}';
+
+  Duration get panelDelay =>
+      gameRef.options.panelDelay ?? const Duration(milliseconds: 300);
+
   @override
   FutureOr<void> onLoad() {
     if (options?.enableHit != false) {
       add(hitBox = vertexShape.hitBox(vertex, this));
     }
     algorithmOnLoad(algorithm);
+    loadOverlay();
     return super.onLoad();
+  }
+
+  void loadOverlay() {
+    var panelBuilder = gameRef.options.vertexPanelBuilder;
+    if (panelBuilder == null) return;
+
+    gameRef.overlays.addEntry(overlayName, (_, game) {
+      return panelBuilder(vertex, gameRef.camera.viewfinder);
+    });
   }
 
   /// Recursively call the onLoad method of the algorithm,
@@ -101,14 +116,14 @@ class VertexComponent extends ShapeComponent
   @override
   void onHoverEnter() {
     graph.hoverVertex = vertex;
-    gameRef.overlays.add('vertex');
+    gameRef.overlays.add(overlayName);
   }
 
   @override
   void onHoverExit() {
     graph.hoverVertex = null;
-    Future.delayed(const Duration(milliseconds: 300), () {
-      gameRef.overlays.remove('vertex');
+    Future.delayed(panelDelay, () {
+      gameRef.overlays.remove(overlayName);
     });
   }